broker: tear down and restart a demoted replica
It's possible for a broker to transition from primary => replica, if it
doesn't happen to observe a keyspace update where it's assignment is
removed prior to it being re-added by the allocator.

If we were the primary, we need to be sure to tear down the replication
pipeline we may be holding. Handle this scenario by cancelling the
current replica and building a new one.

Also add a (hopefully) temporary work-around for the multiple assignment
issue we're currently tracking down.