allocator: ensure we read-through all Etcd revisions before continuing
It was previously possible for some Etcd transactions to be applied
by Checkpoint within a converge() call, and for a failure to then cause
converge() to bubble up an error, which could result in not reading
through the preceeding successful transactions before continuing.

Refactor checkpointTxn to retain the last successful transaction
revision, and ensure we read through this revision regardless of whether
the converge fully completed or failed along the way.