store_azure: if token refresh fails, allow the credential to be invalid
For a period of time until the next retry, which is now controlled by
returning a short interval until the next refresh. It's perfectly fine
if attempts to use the storage pipeline result in errors.

Crucially, don't spin forever attempting to refresh the credential,
as this blocks the current goroutine.