broker: block Appends if fragment listing is too old
Previously, Appends would proceed if there had ever been a successful
fragment listing of the journal. However, users sometimes break or
de-authorize a previously working fragment store, and this means we'll
continue to accept and accumulate data which we're unable to persist.

Now, we'll stop Appends if we haven't completed a listing within the
last four fragment refresh intervals.