fragment: when exiting, Persist falls back to next store on AuthZ failures
Today, fragments are persisted only to the first fragment store in a
listing. However, users sometimes botch the authorization configuration
of a bucket, or disable / delete the bucket, which leaves Gazette in a
awkward position of forever retrying a fragment upload that can never
succeed.

To account for this, introduce a very conservative fallback: if we can
definitively determine that an upload failed due to a authorization
denial or the bucket no longer being use-able, AND we're trying to exit,
then fall back to the next configured fragment store.

We don't want to do this for temporary errors, such as 503s.

To be maximally conservative, we also don't want to do this if we
aren't actively trying to exit.