stores: increase OK health-check interval and jitter
GCS, and possibly other stores, have rate limits on how quickly an
object may be updated that we sometimes see. It succeeds on retry,
but the 5 minute interval when all is okay was motivated by testing,
and is perhaps too aggressive in production.

Also increase jitter to quickly spread out the health checks of blocks
of brokers starting at correlated times.